refactor: migrate ui routes to blueprint
