+++ app/routes.py
@@ -1699,1087 +1699,3 @@ def init_routes(app: Flask) -> None:
             CLOCK_DIGITAL=CLOCK_DIGITAL,
             CLOCK_NAVBAR=CLOCK_NAVBAR,
         )
-
-
-@login_required
-def danger_mode():
-    if request.method == "POST":
-        action = request.form.get("action")
-        if action == "update_shortcut":
-            path_val = request.form.get("shortcut_path")
-            path = Path(path_val) if path_val else None
-            paths, msg = update_chrome_shortcuts_info(path)
-            if paths:
-                joined = ", ".join(str(p) for p in paths)
-                flash(
-                    f"Updated {len(paths)} shortcut{'s' if len(paths) != 1 else ''}: {joined}",
-                    "success",
-                )
-            else:
-                flash(f"Failed to update shortcuts: {msg}", "error")
-        else:
-            enabled = "enabled" in request.form
-            update_setting("DANGER_MODE", "True" if enabled else "False")
-        return redirect(url_for("danger_mode"))
-
-    current = config.get_setting("DANGER_MODE", "True") == "True"
-    templates = template_manager.get_templates()
-    danger_cameras = sorted([name for name, t in templates.items() if t.get("danger")])
-
-    chrome_path = get_chrome_path()
-    shortcut_opts = [
-        str(p) for p in LINUX_PATHS if p.exists() and os.access(p, os.W_OK)
-    ]
-    danger_info = {
-        "browser": os.path.basename(chrome_path) if chrome_path else "N/A",
-        "path": chrome_path or "N/A",
-        "version": (get_chrome_version(chrome_path) if chrome_path else "N/A"),
-        "shortcut": str(first_shortcut_path() or "N/A"),
-        "patched": not shortcuts_need_patch(),
-        "running": is_chrome_debug_port_open("127.0.0.1", config.DANGER_PORT),
-        "port": config.DANGER_PORT,
-    }
-
-    return render_template(
-        "danger.html",
-        enabled=current,
-        danger_cameras=danger_cameras,
-        danger_info=danger_info,
-        shortcut_options=shortcut_opts,
-        page_title="Danger Mode",
-    )
-
-
-def get_active_templates():
-    templates = template_manager.get_templates()
-    active_cameras = []
-    for id, template in templates.items():
-        template.get("name")
-        last_screenshot_time = template.get("last_screenshot_time")
-        if last_screenshot_time:
-            last_update = datetime.strptime(last_screenshot_time, "%Y-%m-%d %H:%M:%S")
-            if last_update > datetime.utcnow() - timedelta(days=1):
-                active_cameras.append(template)
-    return active_cameras
-
-
-@login_required
-def stream():
-    # Get a list of active cameras (with updates within the last 1 day)
-    return render_template("stream.html", page_title="Stream")
-
-
-@login_required
-def get_groups():
-    # Assuming you have a function that returns a list of unique groups
-    groups = get_active_groups()
-    if "all" not in groups:
-        groups = ["all"] + groups
-    return jsonify(groups)
-
-
-@login_required
-def captions():
-    cost_start = request.args.get("cost_start")
-    cost_end = request.args.get("cost_end")
-
-    # Load recent summaries from the database and convert timestamps to ISO
-    entries = []
-    latest_caption = ""
-    try:
-        session_db = SessionLocal()
-        try:
-            records = (
-                session_db.query(Summary)
-                .order_by(Summary.timestamp.desc())
-                .limit(100)
-                .all()
-            )
-            for rec in records:
-                try:
-                    data = json.loads(rec.content)
-                    for ts, text in data.items():
-                        try:
-                            dt = datetime.utcfromtimestamp(int(ts))
-                            iso_ts = dt.strftime("%Y-%m-%dT%H:%M:%SZ")
-                        except Exception:
-                            iso_ts = ts
-                        entries.append({iso_ts: text})
-                except Exception as exc:
-                    logging.error("Failed to parse captions: %s", exc)
-        finally:
-            session_db.close()
-    except Exception:
-        entries = []
-
-    if entries:
-        try:
-            latest_caption = next(iter(entries[0].values()))
-        except Exception:
-            latest_caption = ""
-
-    # Get templates and calculate next capture time
-    templates = template_manager.get_templates()
-    for name, template in templates.items():
-        last_screenshot_time = template.get("last_screenshot_time")
-        frequency = int(
-            template.get("frequency", 30)
-        )  # Default to 30 minutes if not set
-
-        if last_screenshot_time:
-            last_screenshot = datetime.strptime(
-                last_screenshot_time, "%Y-%m-%d %H:%M:%S"
-            )
-            next_screenshot = last_screenshot + timedelta(minutes=frequency)
-            template["next_screenshot_time"] = next_screenshot.strftime(
-                "%Y-%m-%d %H:%M:%S"
-            )
-        else:
-            template["next_screenshot_time"] = None
-
-        templates[name]["screenshot_count"] = template_manager.get_screenshot_count(
-            name
-        )
-        templates[name]["video_count"] = template_manager.get_video_count(name)
-        templates[name]["storage_usage"] = template_manager.get_storage_usage(name)
-        templates[name]["storage_usage_bytes"] = (
-            template_manager.get_storage_usage_bytes(name)
-        )
-        templates[name]["llm_response_count"] = template_manager.get_llm_response_count(
-            name
-        )
-        templates[name]["llm_cost_estimate"] = template_manager.get_llm_cost_estimate(
-            name, start_date=cost_start, end_date=cost_end
-        )
-
-    # Get a list of active cameras (with updates within the last 1 day)
-    return render_template(
-        "captions.html",
-        template_details=templates,
-        lcaptions=entries,
-        latest_caption=latest_caption,
-        page_title="Captions",
-        cost_start=cost_start,
-        cost_end=cost_end,
-    )
-
-
-@login_required
-def download_captions_tsv():
-    """Download all template captions as a TSV file."""
-    templates = template_manager.get_templates()
-
-    # Create a StringIO object to write the TSV data
-    output = io.StringIO()
-    writer = csv.writer(output, delimiter="\t")
-
-    # Write header row
-    writer.writerow(["name", "groups", "notes", "last_caption"])
-
-    # Write data rows
-    for name, template in templates.items():
-        writer.writerow(
-            [
-                name,
-                template.get("groups", ""),
-                template.get("notes", ""),
-                template.get("last_caption", ""),
-            ]
-        )
-
-    # Create response with TSV file
-    output.seek(0)
-    return Response(
-        output.getvalue(),
-        mimetype="text/tab-separated-values",
-        headers={"Content-Disposition": "attachment;filename=captions.tsv"},
-    )
-
-
-@login_required
-def upload_captions_tsv():
-    """Upload and process a TSV file to update template captions."""
-    if "tsv_file" not in request.files:
-        flash("No file part", "error")
-        return redirect(url_for("captions"))
-
-    file = request.files["tsv_file"]
-
-    if file.filename == "":
-        flash("No selected file", "error")
-        return redirect(url_for("captions"))
-
-    if file and file.filename.endswith(".tsv"):
-        # Read the TSV file
-        stream = io.StringIO(file.stream.read().decode("UTF8"), newline=None)
-        reader = csv.reader(stream, delimiter="\t")
-
-        # Skip header row
-        next(reader, None)
-
-        # Process each row
-        updated_count = 0
-        for row in reader:
-            if len(row) >= 4:
-                name, groups, notes, last_caption = row[:4]
-
-                # Validate template name
-                template_name = validate_template_name(name)
-                if template_name is None:
-                    continue
-
-                # Get existing template
-                template = template_manager.get_template(template_name)
-                if template:
-                    # Update template fields
-                    updates = {"groups": groups, "notes": notes}
-
-                    # Only update last_caption if it's different
-                    if last_caption and last_caption != template.get(
-                        "last_caption", ""
-                    ):
-                        updates["last_caption"] = last_caption
-                        updates["last_caption_time"] = datetime.utcnow().strftime(
-                            "%Y-%m-%d %H:%M:%S"
-                        )
-
-                    # Save the updated template
-                    if template_manager.save_template(template_name, updates):
-                        updated_count += 1
-
-        flash(f"Successfully updated {updated_count} templates", "success")
-        return redirect(url_for("captions"))
-
-    flash("Invalid file format. Please upload a TSV file.", "error")
-    return redirect(url_for("captions"))
-
-
-@login_required
-def captions_chat():
-    """Answer a question using recent caption history."""
-
-    data = request.get_json(force=True) or {}
-    question = (data.get("question") or "").strip()
-    if not question:
-        return jsonify({"error": "Missing question"}), 400
-
-    start = data.get("start")
-    end = data.get("end")
-
-    session_db = SessionLocal()
-    try:
-        query = session_db.query(Summary).order_by(Summary.timestamp.desc())
-        if start:
-            try:
-                start_ts = int(datetime.fromisoformat(start).timestamp())
-                query = query.filter(Summary.timestamp >= start_ts)
-            except (ValueError, TypeError) as exc:
-                logging.warning("Invalid start parameter %s: %s", start, exc)
-        if end:
-            try:
-                end_ts = int(datetime.fromisoformat(end).timestamp())
-                query = query.filter(Summary.timestamp <= end_ts)
-            except (ValueError, TypeError) as exc:
-                logging.warning("Invalid end parameter %s: %s", end, exc)
-        records = query.limit(101).all()
-        truncated = len(records) > 100
-        records = records[:100]
-        captions = []
-        for rec in reversed(records):
-            try:
-                jdata = json.loads(rec.content)
-                captions.extend(jdata.values())
-            except Exception:
-                captions.append(rec.content)
-    finally:
-        session_db.close()
-
-    history = "\n".join(captions)
-    answer = ask_question(question, history) or ""
-
-    ts = int(datetime.utcnow().timestamp())
-    session_db = SessionLocal()
-    try:
-        session_db.add(
-            Summary(timestamp=ts, content=json.dumps({ts: f"Q: {question}"}))
-        )
-        if answer:
-            ts2 = ts + 1
-            session_db.add(
-                Summary(
-                    timestamp=ts2,
-                    content=json.dumps({ts2: f"A: {answer}"}),
-                )
-            )
-        session_db.commit()
-    finally:
-        session_db.close()
-
-    return jsonify({"answer": answer, "truncated": truncated})
-
-
-@login_required
-def live():
-    """Render the live view page.
-
-    Optionally filter to a single camera when ``camera`` is provided in the
-    query string. This avoids loading metadata for all cameras when embedding
-    the live view for a specific template.
-    """
-
-    camera = request.args.get("camera")
-    if camera:
-        camera = validate_template_name(camera)
-        if camera is None:
-            abort(400, "Invalid camera name")
-        details = template_manager.get_template(camera)
-        if not details:
-            abort(404)
-        templates = {camera: details}
-    else:
-        templates = template_manager.get_templates()
-
-    return render_template(
-        "live.html", template_details=templates, page_title="Live View"
-    )
-
-
-@login_required
-def clock_page():
-    """Render a standalone clock page."""
-
-    return render_template("clock.html", page_title="Clock")
-
-
-@login_required
-def latest_frame(template_name: TemplateName):
-    """
-    Serve the latest frame for a specific camera.
-    """
-    raw_name = template_name
-    template_name = validate_template_name(str(template_name))
-    if template_name is None:
-        logging.warning("Unable to serve screenshot for %s", raw_name)
-        resp = send_conditional_file(
-            _placeholder_screenshot(),
-            cache_seconds=PNG_TTL_SEC,
-            mimetype="image/png",
-        )
-        resp.status_code = 404
-        return resp
-
-    path = os.path.join(
-        os.path.dirname(os.path.join(__file__)),
-        "..",
-        SCREENSHOT_DIRECTORY,
-        str(template_name),
-    )
-    if not os.path.exists(path):
-        abort(404)
-
-    latest_file = max(
-        (f for f in os.listdir(path) if f.endswith(".png")),
-        key=lambda f: os.path.getmtime(os.path.join(path, f)),
-    )
-
-    if latest_file:
-        return send_conditional_file(os.path.join(path, latest_file), PNG_TTL_SEC)
-
-    abort(404)
-
-
-@login_required
-def upload_nav_icon():
-    """Upload or choose a navigation logo."""
-
-    choice = request.form.get("logo_choice")
-    if choice in {"img/glimpser_small.png", "img/glimpser.png"}:
-        update_setting("NAV_ICON", choice)
-        flash("Navigation logo updated", "success")
-        return redirect(url_for("ui.settings"))
-
-    if "logo_file" not in request.files:
-        flash("No logo file provided", "error")
-        return redirect(url_for("ui.settings")), 400
-
-    logo_file = request.files["logo_file"]
-    if logo_file.filename == "":
-        flash("No logo file provided", "error")
-        return redirect(url_for("ui.settings")), 400
-
-    if not allowed_filename(
-        logo_file.filename
-    ) or not logo_file.filename.lower().endswith(".png"):
-        flash("Invalid file name", "error")
-        return redirect(url_for("ui.settings")), 400
-
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        logo_file.save(temp_file.name)
-        try:
-            img = Image.open(temp_file.name)
-            w, h = img.size
-            if h == 0 or not 2 <= w / h <= 10:
-                os.unlink(temp_file.name)
-                flash("Invalid aspect ratio", "error")
-                return redirect(url_for("ui.settings")), 400
-        except Exception:
-            os.unlink(temp_file.name)
-            flash("Invalid image file", "error")
-            return redirect(url_for("ui.settings")), 400
-
-        dest_dir = os.path.join(current_app.static_folder, "img")
-        os.makedirs(dest_dir, exist_ok=True)
-        dest_name = secure_filename(logo_file.filename)
-        dest_path = os.path.join(dest_dir, dest_name)
-        shutil.move(temp_file.name, dest_path)
-        update_setting("NAV_ICON", f"img/{dest_name}")
-    flash("Navigation logo uploaded", "success")
-    return redirect(url_for("ui.settings"))
-
-
-@login_required
-def manage_templates():
-    if request.method == "POST":
-        data = request.get_json(silent=True) or {}
-        template_name = validate_template_name(str(data.get("name", "")))
-        if template_name is None:
-            abort(404)
-        url = data.get("url", "")
-        if url and ("onvif" in url or urlparse(url).path in {"", "/"}):
-            try:
-                endpoints = camera_discovery.autodetect_onvif_endpoints(url)
-                if endpoints.get("snapshot"):
-                    data["url"] = endpoints["snapshot"]
-                elif endpoints.get("stream"):
-                    data["url"] = endpoints["stream"]
-            except Exception:
-                pass
-        if template_manager.save_template(template_name, data):
-            return jsonify({"status": "success", "message": "Template saved"})
-
-    elif request.method == "GET":
-        group = request.args.get("group")
-        search_query = request.args.get("search", "").lower()
-        templates = template_manager.get_templates()
-
-        filtered_templates = {}
-        for name, template in templates.items():
-            template_groups = template.get("groups", "").split(",")
-            if (group == "all" or group in template_groups) and (
-                not search_query
-                or search_query in name.lower()
-                or search_query in template.get("url", "").lower()
-                or any(search_query in g.lower() for g in template_groups)
-            ):
-                filtered_templates[name] = template
-
-        return jsonify(filtered_templates)
-
-    elif request.method == "DELETE":
-        data = request.get_json(silent=True) or {}
-        template_name = validate_template_name(str(data.get("name", "")))
-        if template_name is None:
-            abort(404)
-        if template_manager.delete_template(template_name):
-            return jsonify({"status": "success", "message": "Template deleted"})
-        else:
-            return (
-                jsonify({"status": "failure", "message": "Template not found"}),
-                404,
-            )
-
-
-@login_required
-def template_details(template_name: TemplateName):
-    template_name = validate_template_name(str(template_name))
-    if template_name is None:
-        abort(404)
-
-    templates = template_manager.get_templates()
-    template_details = templates.get(template_name)
-    if template_details is None:
-        abort(404)  # Template not found
-    lscreenshots = template_manager.get_screenshots_for_template(template_name)
-    lvideos = template_manager.get_videos_for_template(template_name)
-    object_tokens = ["person", "car", "dog", "cat", "truck", "bus", "bicycle"]
-    return render_template(
-        "template_details.html",
-        template_name=template_name,
-        template_details=template_details,
-        screenshots=lscreenshots,
-        videos=lvideos,
-        object_tokens=object_tokens,
-        clip_model=CLIP_MODEL_NAME,
-        clip_gpu=clip_gpu_available(),
-        page_title="Camera Details",
-    )
-
-
-@login_required
-def generate_prompt_route(template_name: TemplateName):
-    """Return a suggested caption prompt for ``template_name``."""
-
-    template_name = validate_template_name(str(template_name))
-    if template_name is None:
-        abort(404)
-
-    prompt = prompt_optimizer.generate_prompt(template_name)
-    return jsonify({"prompt": prompt})
-
-
-@login_required
-def suggest_fix_route(template_name: TemplateName):
-    """Return diagnostic info and replacement URL suggestions."""
-
-    template_name = validate_template_name(str(template_name))
-    if template_name is None:
-        abort(404)
-
-    details = template_manager.get_template(template_name)
-    if not details:
-        abort(404)
-
-    url = details.get("url", "")
-    xpaths = [
-        details.get("popup_xpath", ""),
-        details.get("dedicated_xpath", ""),
-    ]
-    info = camera_fix.check_camera_template(url, xpaths)
-    return jsonify(info)
-
-
-@login_required
-def camera_diagnostics(template_name: TemplateName) -> Response:
-    """Return live diagnostics for ``template_name``.
-
-    The diagnostics include ping latency, open ports, HTTP banner
-    information and a simple device type label when available.
-    """
-
-    template_name = validate_template_name(str(template_name))
-    if template_name is None:
-        abort(404)
-
-    details = template_manager.get_template(template_name)
-    if not details:
-        abort(404)
-
-    url = details.get("url", "")
-    host = urlparse(url).hostname or url
-    try:
-        ip = socket.gethostbyname(host)
-    except Exception:
-        ip = host
-
-    data: dict[str, typing.Any] = {"ip": ip}
-
-    latency = camera_discovery._ping_latency(ip)
-    if latency is not None:
-        data["ping_ms"] = latency
-
-    ports = camera_discovery._detect_open_ports(ip, camera_discovery.COMMON_PORTS)
-    if ports:
-        data["open_ports"] = ports
-        for p in ports:
-            if p in (80, 8080, 443):
-                banner = camera_discovery._fetch_http_banner(ip, p)
-                for k, v in banner.items():
-                    data.setdefault(k, v)
-
-    dtype = camera_discovery._classify_device(
-        {
-            "ip": ip,
-            "protocol": details.get("protocol", urlparse(url).scheme or "http"),
-            "port": details.get("port", urlparse(url).port or 0),
-            "info": data,
-        }
-    )
-    if dtype:
-        data["device_type"] = dtype
-
-    return jsonify(data)
-
-
-def delete_setting(name: str) -> bool:
-    name = name.replace("'", "")[:32]
-    if not re.findall(r"^[A-Z_]+?$", name):
-        return False
-
-    session = SessionLocal()
-    try:
-        session.execute(text("DELETE FROM settings WHERE name = :name"), {"name": name})
-        session.commit()
-    finally:
-        session.close()
-
-    return True
-
-
-@login_required
-def settings():
-    """Render settings page and handle configuration updates via POST."""
-    if request.method == "POST":
-        email_settings = [
-            "EMAIL_ENABLED",
-            "EMAIL_SENDER",
-            "EMAIL_RECIPIENTS",
-            "EMAIL_SMTP_SERVER",
-            "EMAIL_SMTP_PORT",
-            "EMAIL_USE_TLS",
-            "EMAIL_USERNAME",
-            "EMAIL_PASSWORD",
-        ]
-        notification_settings = ["SMS_ENABLED", "CAP_ENABLED"]
-        action = request.form.get("action")
-        if action == "add":
-            new_name = (request.form.get("new_name") or "").strip()
-            new_value = (request.form.get("new_value") or "").strip()
-            if not new_name or not new_value:
-                flash("Setting name and value are required", "error")
-                return redirect(url_for("ui.settings")), 400
-            if not re.fullmatch(r"[A-Z_]+", new_name):
-                flash(
-                    "Setting names must contain only uppercase letters and underscores",
-                    "error",
-                )
-                return redirect(url_for("ui.settings")), 400
-            sanitized = validate_setting(new_name, new_value)
-            if sanitized is None:
-                flash(f"Invalid value for {new_name}", "error")
-                return redirect(url_for("ui.settings")), 400
-            update_setting(new_name, sanitized)
-        elif action == "delete":
-            name_to_delete = request.form.get("name_to_delete")
-            if name_to_delete:
-                delete_setting(name_to_delete)
-        elif action == "update_email":
-            for setting in email_settings:
-                value = request.form.get(setting)
-                if value is not None:
-                    sanitized = validate_setting(setting, value)
-                    if sanitized is None:
-                        flash(f"Invalid value for {setting}", "error")
-                        return redirect(url_for("ui.settings")), 400
-                    update_setting(setting, sanitized)
-        elif action == "backup":
-            if backup_config():
-                flash("Configuration backed up successfully", "success")
-            else:
-                flash("Failed to backup configuration", "error")
-        elif action == "download":
-
-            def generate() -> typing.Generator[bytes, None, None]:
-                """Stream the configuration JSON after triggering a backup."""
-                yield b""
-                backup_config()
-                if os.path.exists(BACKUP_PATH):
-                    with open(BACKUP_PATH, "rb") as f:
-                        for chunk in iter(lambda: f.read(8192), b""):
-                            yield chunk
-
-            headers = {"Content-Disposition": "attachment; filename=config_backup.json"}
-            return Response(
-                stream_with_context(generate()),
-                mimetype="application/json",
-                headers=headers,
-            )
-        elif action == "upload":
-            if "file" not in request.files:
-                flash("No file part", "error")
-            else:
-                file = request.files["file"]
-                if file.filename == "":
-                    flash("No selected file", "error")
-                elif file and allowed_file(file.filename):
-                    file.stream.seek(0, os.SEEK_END)
-                    size = file.stream.tell()
-                    file.stream.seek(0)
-                    if size > MAX_UPLOAD_SIZE:
-                        flash("File exceeds 5 MB limit", "error")
-                    else:
-                        file.save(BACKUP_PATH)
-                        restore_config()
-                        flash(
-                            "Configuration restored successfully",
-                            "success",
-                        )
-                else:
-                    flash("Invalid file type", "error")
-        elif action == "test_email":
-            send_email_alert(
-                "Glimpser Test Email",
-                "This is a test email from Glimpser.",
-            )
-            flash("Email test triggered. Check logs for results.", "info")
-        elif action == "test_sms":
-            send_sms_alert("Test SMS from Glimpser")
-            flash("SMS test triggered. Check logs for results.", "info")
-        elif action == "update_shortcut":
-            path_val = request.form.get("shortcut_path")
-            path = Path(path_val) if path_val else None
-            paths, msg = update_chrome_shortcuts_info(path)
-            if paths:
-                joined = ", ".join(str(p) for p in paths)
-                flash(
-                    f"Updated {len(paths)} shortcut{'s' if len(paths) != 1 else ''}: {joined}",
-                    "success",
-                )
-            else:
-                flash(f"Failed to update shortcuts: {msg}", "error")
-        else:
-            current = {s["name"]: s["value"] for s in get_all_settings()}
-            bool_settings = {
-                n for n, v in current.items() if validators.is_bool_string(v)
-            }
-            for name in bool_settings:
-                if name in email_settings:
-                    continue
-                new_val = (
-                    "True"
-                    if str(request.form.get(name, "")).lower()
-                    in {"true", "on", "1", "t", "y", "yes"}
-                    else "False"
-                )
-                update_setting(name, new_val)
-
-            for name, value in request.form.items():
-                if (
-                    name
-                    in [
-                        "action",
-                        "new_name",
-                        "new_value",
-                        "name_to_delete",
-                    ]
-                    or name in email_settings
-                    or name in bool_settings
-                ):
-                    continue
-
-                sanitized = validate_setting(name, value)
-                if sanitized is None:
-                    flash(f"Invalid value for {name}", "error")
-                    continue
-
-                if name in SETTINGS_CHOICES:
-                    update_setting(name, sanitized)
-                    continue
-
-                update_setting(name, sanitized)
-        flash("Settings updated successfully", "success")
-        return redirect(url_for("ui.settings"))
-
-    settings = get_all_settings()
-    grouped_settings: Dict[str, List[Dict[str, Any]]] = {
-        group: [] for group in SETTINGS_GROUPS
-    }
-    grouped_settings["Other"] = []
-    for setting in settings:
-        placed = False
-        for group, names in SETTINGS_GROUPS.items():
-            if setting["name"] in names:
-                grouped_settings[group].append(setting)
-                placed = True
-                break
-        if not placed:
-            grouped_settings["Other"].append(setting)
-
-    if "Advanced" in grouped_settings:
-        grouped_settings["Advanced"].extend(grouped_settings.get("Other", []))
-    else:
-        grouped_settings["Advanced"] = grouped_settings.get("Other", [])
-    grouped_settings.pop("Other", None)
-
-    # Remove empty groups to avoid blank headings in the UI
-    grouped_settings = {g: items for g, items in grouped_settings.items() if items}
-
-    collapsed_groups = {
-        "Capture",
-        "Admin",
-        "Advanced",
-    }
-    file_location_items = [s for s in settings if s["name"] in FILE_LOCATION_NAMES]
-    file_info = file_location_metrics(file_location_items)
-
-    metrics = scheduling.get_system_metrics()
-    feeds = scheduling.get_feed_status()
-    last_summary = scheduling.get_last_summary_time()
-    log_summary = scheduling.get_or_generate_log_summary()
-    danger_enabled = config.get_setting("DANGER_MODE", "True") == "True"
-    cost_summary, total_tokens, total_cost, total_calls = (
-        template_manager.get_llm_cost_summary()
-    )
-    cost_summary = template_manager.group_cost_summary(cost_summary, top=10)
-
-    chrome_path = get_chrome_path()
-    shortcut_opts = [
-        str(p) for p in LINUX_PATHS if p.exists() and os.access(p, os.W_OK)
-    ]
-    danger_info = {
-        "browser": os.path.basename(chrome_path) if chrome_path else "N/A",
-        "path": chrome_path or "N/A",
-        "version": (get_chrome_version(chrome_path) if chrome_path else "N/A"),
-        "shortcut": str(first_shortcut_path() or "N/A"),
-        "patched": not shortcuts_need_patch(),
-        "running": is_chrome_debug_port_open("127.0.0.1", config.DANGER_PORT),
-        "port": config.DANGER_PORT,
-    }
-
-    last_backup = None
-    if os.path.exists(BACKUP_PATH):
-        ts = datetime.fromtimestamp(os.path.getmtime(BACKUP_PATH))
-        last_backup = ts.strftime("%Y-%m-%d %H:%M:%S")
-
-    templates = template_manager.get_templates()
-    existing_urls = {t.get("url"): n for n, t in templates.items() if t.get("url")}
-
-    tooltips = dict(SETTINGS_TOOLTIPS)
-    if not metrics.get("ffmpeg_gpu_support"):
-        tooltips["FFMPEG_HWACCEL"] = "Hardware acceleration not available"
-
-    return render_template(
-        "settings.html",
-        grouped_settings=grouped_settings,
-        collapsed_groups=collapsed_groups,
-        tooltips=tooltips,
-        metrics=metrics,
-        feeds=feeds,
-        last_summary=last_summary,
-        log_summary=log_summary,
-        cost_summary=cost_summary,
-        total_tokens=total_tokens,
-        total_cost=total_cost,
-        total_calls=total_calls,
-        danger_info=danger_info,
-        shortcut_options=shortcut_opts,
-        choices=SETTINGS_CHOICES,
-        boolean_fields=validators.BOOLEAN_SETTINGS,
-        danger_enabled=danger_enabled,
-        numeric_fields=NUMERIC_FIELDS,
-        email_fields=EMAIL_FIELDS,
-        locked_settings=LOCKED_SETTINGS,
-        file_info=file_info,
-        existing_urls=existing_urls,
-        placeholders=SETTINGS_PLACEHOLDERS,
-        last_backup=last_backup,
-        page_title="Settings",
-    )
-
-
-def allowed_file(filename):
-    return "." in filename and filename.rsplit(".", 1)[1].lower() == "json"
-
-
-@login_required
-def update_template(template_name: TemplateName):
-    template_name = validate_template_name(str(template_name))
-    if template_name is None:
-        abort(404)
-
-    if True:
-        # Extract form data
-
-        updated_data = {
-            "url": request.form.get("url"),
-            "frequency": request.form.get("frequency"),
-            "timeout": request.form.get("timeout"),
-            "notes": request.form.get("notes"),
-            "popup_xpath": request.form.get("popup_xpath"),
-            "dedicated_xpath": request.form.get("dedicated_xpath"),
-            "callback_url": request.form.get("callback_url"),
-            "proxy": request.form.get("proxy"),
-            "auth_username": request.form.get("auth_username"),
-            "auth_password": request.form.get("auth_password"),
-            "rollback_frames": request.form.get("rollback_frames"),
-            "groups": request.form.get("groups"),
-            "object_filter": request.form.get("object_filter"),
-            "object_confidence": request.form.get("object_confidence", 0.5),
-            "motion": request.form.get("motion", 0.2),
-            "invert": request.form.get("invert", "false").lower()
-            in ["true", "1", "t", "y", "yes", "on"],
-            "dark": request.form.get("dark", "false").lower()
-            in ["true", "1", "t", "y", "yes", "on"],
-            "headless": request.form.get("headless", "false").lower()
-            in ["true", "1", "t", "y", "yes", "on"],
-            "stealth": request.form.get("stealth", "false").lower()
-            in ["true", "1", "t", "y", "yes", "on"],
-            "browser": request.form.get("browser", "false").lower()
-            in ["true", "1", "t", "y", "yes", "on"],
-            "livecaption": request.form.get("livecaption", "false").lower()
-            in ["true", "1", "t", "y", "yes", "on"],
-            "danger": request.form.get("danger", "false").lower()
-            in ["true", "1", "t", "y", "yes", "on"],
-        }
-
-        lremoves = []
-        for lkey in updated_data:
-            if updated_data.get(lkey) is None:
-                lremoves.append(lkey)
-        for lkey in lremoves:
-            del updated_data[lkey]
-
-        # Validate and normalize the incoming form data. Any missing or
-        # out-of-range values are replaced with sensible defaults. A
-        # ``ValueError`` is raised when required fields are absent.
-        try:
-            updated_data = validate_update_data(updated_data)
-        except ValueError as exc:
-            if request.is_json:
-                return jsonify({"error": str(exc)}), 400
-            flash(str(exc), "error")
-            return redirect("/templates/" + template_name)
-
-        # Update the template in your storage (e.g., JSON file, database)
-        # This assumes you have a function to update templates
-        template_manager.save_template(template_name, updated_data)
-
-        # Stop the old scheduled job and generate a blank frame so the
-        # updated template is clearly segmented in any compiled video.
-        try:
-            scheduling.scheduler.remove_job(template_name)
-        except Exception:
-            pass
-        screenshots.create_blank_frame(template_name)
-        template_manager.get_template(template_name)
-        try:
-            seconds = int(updated_data.get("frequency", 30 * 60))
-            scheduling.scheduler.add_job(
-                func=scheduling.update_camera,
-                trigger="interval",
-                seconds=seconds,
-                args=[template_name, updated_data],
-                id=template_name,
-                replace_existing=True,
-            )
-        except Exception as e:
-            logging.error("job schedule error: %s", e)
-
-        if request.is_json:
-            return jsonify({"message": "Template updated successfully!"})
-
-        return redirect("/templates/" + template_name)
-
-
-@login_required
-def status():
-    """Redirect to the Status tab under Settings for consistency."""
-    return redirect(url_for("ui.settings", tab="status-tab"))
-
-
-@login_required
-def logs():
-    return render_template("logs.html", page_title="Logs")
-
-
-@login_required
-def cost_summary_page():
-    templates = template_manager.get_templates()
-    costs = {name: template_manager.get_llm_cost_estimate(name) for name in templates}
-    start_time = int(scheduling.system_metrics.get("start_time", time.time()))
-    return render_template(
-        "cost_summary.html",
-        costs=costs,
-        start_time=start_time,
-        page_title="LLM Cost Summary",
-    )
-
-
-@login_required
-def api_llm_cost_summary():
-    start = request.args.get("start")
-    end = request.args.get("end")
-    group = request.args.get("group")
-    summary, _, _, _ = template_manager.get_llm_cost_summary(
-        start_date=start, end_date=end, group=group
-    )
-    costs = {
-        entry["name"]: {
-            "tokens": entry["tokens"],
-            "cost": entry["cost"],
-            "calls": entry.get("calls", 0),
-        }
-        for entry in summary
-    }
-    return jsonify(costs)
-
-
-@login_required
-def api_camera_log_summary(template_name: TemplateName):
-    template_name = validate_template_name(str(template_name))
-    if template_name is None:
-        abort(404)
-    summary = scheduling.get_or_generate_camera_log_summary(template_name)
-    if summary is None:
-        return ("", 204)
-    return jsonify({"summary": summary})
-
-
-@login_required
-def stream_logs():
-
-    level = request.args.get("level")
-    source = request.args.get("source")
-    start_date = request.args.get("start_date")
-    end_date = request.args.get("end_date")
-    search = request.args.get("search")
-
-    user_id = session.get("user_id", 0)
-    combo = (int(user_id), level or "", search or "")
-    if combo in active_log_streams:
-        # avoid spawning duplicate streams for the same parameters
-        return Response(
-            "event: duplicate\ndata: {}\n\n",
-            mimetype="text/event-stream",
-        )
-    active_log_streams[combo] = True
-
-    def generate():
-        try:
-            while True:
-                # Read and filter logs from memory
-                logs = read_logs_from_memory(
-                    level=level,
-                    source=source,
-                    start_date=start_date,
-                    end_date=end_date,
-                    search=search,
-                )
-
-                # Limit the number of logs sent to improve performance
-                logs = logs[:50]
-
-                yield f"data: {json.dumps(logs, default=str)}\n\n"
-                time.sleep(1)  # Send updates every second
-        finally:
-            active_log_streams.pop(combo, None)
-
-    return Response(
-        stream_with_context(generate()),
-        mimetype="text/event-stream",
-    )
-
-
-@login_required
-def search_suggestions():
-    """Return autocomplete suggestions for the search box."""
-
-    query = request.args.get("q", "").lower()
-    templates = template_manager.get_templates()
-    names = [t.get("name", "") for t in templates.values()]
-    groups = get_active_groups()
-
-    docs_path = Path(__file__).resolve().parent.parent / "docs"
-    doc_names = [
-        f.stem for f in docs_path.glob("*.md") if f.name.lower() != "readme.md"
-    ]
-
-    suggestions: list[str] = []
-    for item in names + groups + doc_names:
-        if query and query not in item.lower():
-            continue
-        if item not in suggestions:
-            suggestions.append(item)
-        if len(suggestions) >= 10:
-            break
-
-    return jsonify(suggestions)
